Rename suites to monitoring for better organization
