+++ monitoring/functional/sync.test.ts
