+++ monitoring/networkchaos/group-reconciliation.test.ts
@@ -3,7 +3,7 @@ import { setupDurationTracking } from "@helpers/vitest";
 import { getWorkers } from "@workers/manager";
 import type { Group } from "@workers/versions";
 import { describe, expect, it } from "vitest";
-import { DockerContainer } from "../../network-stability-utilities/container";
+import { DockerContainer } from "../../network-stability/container";
 
 const testName = "group-reconciliation";
 describe(testName, async () => {
