+++ monitoring/networkchaos/keyrotation.test.ts
@@ -3,7 +3,7 @@ import { setupDurationTracking } from "@helpers/vitest";
 import { typeofStream } from "@workers/main";
 import { getWorkers } from "@workers/manager";
 import { describe, expect, it } from "vitest";
-import { DockerContainer } from "../../network-stability-utilities/container";
+import { DockerContainer } from "../../network-stability/container";
 
 const testName = "keyrotation-chaos";
 describe(testName, async () => {
