--- network-stability-utilities/README.txt
@@ -1,2 +1,0 @@
-Must have Docker installed and `docker` command available to run as the test user
-Must have passwordless SUDO enabled
